Make the showcase image setup more responsive
The general scheme is that there is an image that is as wide as the page
and the content is absolutely positioned on the page below it. There is
javascript in place that adjust the top of the content on load and on
window resize to make the very top of the content block visible.

The adjusting of the content block is required so that people are more
adeqately aware that there is content to scroll to.