Add a loading screen
Sometimes the pages take a while to load and until they do load the
layout is extremely malformed. Optimally the page load times would
simply be improved. Unfortunately I am having some difficulty doing that
so this is the next best thing. At least if there is a loading image the
user knows that the page is still being loaded.