Update photos and make visual adjustments better
Taylor wanted to swap out some of the photos for the showcases.

Also I have adjusted the mechanism that helps keep the right parts of
the photo visible to be a little better at doing its job. Rather than
just taking space from the bottom of the photo space is reclaimed from
both the top and bottom.